remove set_pat from edit_in_place and move set_pat_with_editor in edit
